refactor(storage): refactor get_from_table
