+++ src/storage/src/hummock/mod.rs
@@ -66,9 +66,12 @@ use self::key::user_key;
 pub use self::sstable_store::*;
 pub use self::state_store::HummockStateStoreIter;
 use super::monitor::StateStoreMetrics;
+use crate::error::StorageResult;
 use crate::hummock::compaction_group_client::{CompactionGroupClient, DummyCompactionGroupClient};
 use crate::hummock::conflict_detector::ConflictDetector;
 use crate::hummock::local_version_manager::LocalVersionManager;
+use crate::hummock::shared_buffer::shared_buffer_batch::SharedBufferBatch;
+use crate::hummock::shared_buffer::{OrderSortedUncommittedData, UncommittedData};
 use crate::hummock::sstable::SstableIteratorReadOptions;
 use crate::hummock::sstable_store::{SstableStoreRef, TableHolder};
 use crate::monitor::StoreLocalStatistic;
@@ -162,11 +165,11 @@ impl HummockStorage {
         &self,
         sstable: TableHolder,
         internal_key: &[u8],
-        key: &[u8],
         check_bloom_filter: bool,
         stats: &mut StoreLocalStatistic,
     ) -> HummockResult<Option<Option<Bytes>>> {
-        if check_bloom_filter && !Self::hit_sstable_bloom_filter(sstable.value(), key, stats) {
+        let ukey = user_key(internal_key);
+        if check_bloom_filter && !Self::hit_sstable_bloom_filter(sstable.value(), ukey, stats) {
             return Ok(None);
         }
 
@@ -185,7 +188,7 @@ impl HummockStorage {
 
         // Iterator gets us the key, we tell if it's the key we want
         // or key next to it.
-        let value = match user_key(iter.key()) == key {
+        let value = match key::user_key(iter.key()) == ukey {
             true => Some(iter.value().into_user_value().map(Bytes::copy_from_slice)),
             false => None,
         };
@@ -232,5 +235,51 @@ impl HummockStorage {
         }
 
         !surely_not_have
+    }
+
+    /// Get `user_value` from `OrderSortedUncommittedData`. If not get successful, return None.
+    async fn get_from_order_sorted_uncommitted_data(
+        &self,
+        order_sorted_uncommitted_data: OrderSortedUncommittedData,
+        internal_key: &[u8],
+        stats: &mut StoreLocalStatistic,
+        key: &[u8],
+        check_bloom_filter: bool,
+    ) -> StorageResult<(Option<Option<Bytes>>, i32)> {
+        let mut table_counts = 0;
+        let epoch = key::get_epoch(internal_key);
+        for data_list in order_sorted_uncommitted_data {
+            for data in data_list {
+                match data {
+                    UncommittedData::Batch(batch) => {
+                        assert!(batch.epoch() <= epoch, "batch'epoch greater than epoch");
+                        let data = self.get_from_batch(&batch, key);
+                        if data.is_some() {
+                            return Ok((data, table_counts));
+                        }
+                    }
+                    UncommittedData::Sst((_, table_info)) => {
+                        let table = self.sstable_store.sstable(table_info.id, stats).await?;
+                        table_counts += 1;
+
+                        let data = self
+                            .get_from_table(table, internal_key, check_bloom_filter, stats)
+                            .await?;
+                        if data.is_some() {
+                            return Ok((data, table_counts));
+                        }
+                    }
+                }
+            }
+        }
+        Ok((None, table_counts))
+    }
+
+    /// Get `user_value` from `SharedBufferBatch` Return `Some(None)` means the key is deleted.
+    fn get_from_batch(&self, batch: &SharedBufferBatch, key: &[u8]) -> Option<Option<Bytes>> {
+        batch.get(key).map(|v| {
+            self.stats.get_shared_buffer_hit_counts.inc();
+            v.into_user_value().map(|v| v.into())
+        })
     }
 }
